+++ src/executor.ts
@@ -13,6 +13,7 @@ import {spawn} from 'child_process';
 import {WireitError} from './error.js';
 import {scriptReferenceToString} from './script.js';
 import {shuffle} from './util/shuffle.js';
+import {WorkerPool} from './util/worker-pool.js';
 
 import type {
   ScriptConfig,
@@ -40,16 +41,23 @@ const IS_WINDOWS = process.platform === 'win32';
 export class Executor {
   readonly #cache = new Map<string, Promise<CacheKey | typeof UNCACHEABLE>>();
   readonly #logger: Logger;
+  readonly #workerPool: WorkerPool;
 
-  constructor(logger: Logger) {
+  constructor(logger: Logger, workerPool: WorkerPool) {
     this.#logger = logger;
+    this.#workerPool = workerPool;
   }
 
   async execute(script: ScriptConfig): Promise<CacheKey | typeof UNCACHEABLE> {
     const cacheKey = scriptReferenceToString(script);
     let promise = this.#cache.get(cacheKey);
     if (promise === undefined) {
-      promise = ScriptExecution.execute(script, this, this.#logger);
+      promise = ScriptExecution.execute(
+        script,
+        this,
+        this.#workerPool,
+        this.#logger
+      );
       this.#cache.set(cacheKey, promise);
     }
     return promise;
@@ -63,22 +71,26 @@ class ScriptExecution {
   static execute(
     script: ScriptConfig,
     executor: Executor,
+    workerPool: WorkerPool,
     logger: Logger
   ): Promise<CacheKey | typeof UNCACHEABLE> {
-    return new ScriptExecution(script, executor, logger).#execute();
+    return new ScriptExecution(script, executor, workerPool, logger).#execute();
   }
 
   readonly #script: ScriptConfig;
-  readonly #logger: Logger;
   readonly #executor: Executor;
+  readonly #workerPool: WorkerPool;
+  readonly #logger: Logger;
 
   private constructor(
     script: ScriptConfig,
     executor: Executor,
+    workerPool: WorkerPool,
     logger: Logger
   ) {
     this.#script = script;
     this.#executor = executor;
+    this.#workerPool = workerPool;
     this.#logger = logger;
   }
 
@@ -186,104 +198,107 @@ class ScriptExecution {
       detail: 'running',
     });
 
-    // TODO(aomarks) Fix PATH and npm_ environment variables to reflect the new
-    // package when cross-package dependencies are supported.
+    const command = this.#script.command;
+    await this.#workerPool.run(async () => {
+      // TODO(aomarks) Fix PATH and npm_ environment variables to reflect the new
+      // package when cross-package dependencies are supported.
 
-    const child = spawn(this.#script.command, {
-      cwd: this.#script.packageDir,
-      // Conveniently, "shell:true" has the same shell-selection behavior as
-      // "npm run", where on macOS and Linux it is "sh", and on Windows it is
-      // %COMSPEC% || "cmd.exe".
-      //
-      // References:
-      //   https://nodejs.org/api/child_process.html#child_processspawncommand-args-options
-      //   https://nodejs.org/api/child_process.html#default-windows-shell
-      //   https://github.com/npm/run-script/blob/a5b03bdfc3a499bf7587d7414d5ea712888bfe93/lib/make-spawn-args.js#L11
-      shell: true,
-    });
-
-    // Only create the stdout/stderr replay files if we encounter anything on
-    // this streams.
-    let stdoutReplay: WriteStream | undefined;
-    let stderrReplay: WriteStream | undefined;
-
-    child.stdout.on('data', (data: string | Buffer) => {
-      this.#logger.log({
-        script: this.#script,
-        type: 'output',
-        stream: 'stdout',
-        data,
-      });
-      if (stdoutReplay === undefined) {
-        stdoutReplay = createWriteStream(this.#stdoutReplayPath);
-      }
-      stdoutReplay.write(data);
-    });
-
-    child.stderr.on('data', (data: string | Buffer) => {
-      this.#logger.log({
-        script: this.#script,
-        type: 'output',
-        stream: 'stderr',
-        data,
-      });
-      if (stderrReplay === undefined) {
-        stderrReplay = createWriteStream(this.#stderrReplayPath);
-      }
-      stderrReplay.write(data);
-    });
-
-    const completed = new Promise<void>((resolve, reject) => {
-      child.on('error', (error) => {
-        reject(
-          new WireitError({
-            script: this.#script,
-            type: 'failure',
-            reason: 'spawn-error',
-            message: error.message,
-          })
-        );
+      const child = spawn(command, {
+        cwd: this.#script.packageDir,
+        // Conveniently, "shell:true" has the same shell-selection behavior as
+        // "npm run", where on macOS and Linux it is "sh", and on Windows it is
+        // %COMSPEC% || "cmd.exe".
+        //
+        // References:
+        //   https://nodejs.org/api/child_process.html#child_processspawncommand-args-options
+        //   https://nodejs.org/api/child_process.html#default-windows-shell
+        //   https://github.com/npm/run-script/blob/a5b03bdfc3a499bf7587d7414d5ea712888bfe93/lib/make-spawn-args.js#L11
+        shell: true,
       });
 
-      child.on('close', (status, signal) => {
-        if (signal !== null) {
-          reject(
-            new WireitError({
-              script: this.#script,
-              type: 'failure',
-              reason: 'signal',
-              signal,
-            })
-          );
-        } else if (status !== 0) {
-          reject(
-            new WireitError({
-              script: this.#script,
-              type: 'failure',
-              reason: 'exit-non-zero',
-              // status should only ever be null if signal was not null, but
-              // this isn't reflected in the TypeScript types. Just in case, and
-              // to make TypeScript happy, fall back to -1 (which is a
-              // conventional exit status used for "exited with signal").
-              status: status ?? -1,
-            })
-          );
-        } else {
-          resolve();
+      // Only create the stdout/stderr replay files if we encounter anything on
+      // this streams.
+      let stdoutReplay: WriteStream | undefined;
+      let stderrReplay: WriteStream | undefined;
+
+      child.stdout.on('data', (data: string | Buffer) => {
+        this.#logger.log({
+          script: this.#script,
+          type: 'output',
+          stream: 'stdout',
+          data,
+        });
+        if (stdoutReplay === undefined) {
+          stdoutReplay = createWriteStream(this.#stdoutReplayPath);
         }
+        stdoutReplay.write(data);
       });
-    });
 
-    try {
-      await completed;
-    } finally {
-      if (stdoutReplay !== undefined) {
-        await closeWriteStream(stdoutReplay);
+      child.stderr.on('data', (data: string | Buffer) => {
+        this.#logger.log({
+          script: this.#script,
+          type: 'output',
+          stream: 'stderr',
+          data,
+        });
+        if (stderrReplay === undefined) {
+          stderrReplay = createWriteStream(this.#stderrReplayPath);
+        }
+        stderrReplay.write(data);
+      });
+
+      const completed = new Promise<void>((resolve, reject) => {
+        child.on('error', (error) => {
+          reject(
+            new WireitError({
+              script: this.#script,
+              type: 'failure',
+              reason: 'spawn-error',
+              message: error.message,
+            })
+          );
+        });
+
+        child.on('close', (status, signal) => {
+          if (signal !== null) {
+            reject(
+              new WireitError({
+                script: this.#script,
+                type: 'failure',
+                reason: 'signal',
+                signal,
+              })
+            );
+          } else if (status !== 0) {
+            reject(
+              new WireitError({
+                script: this.#script,
+                type: 'failure',
+                reason: 'exit-non-zero',
+                // status should only ever be null if signal was not null, but
+                // this isn't reflected in the TypeScript types. Just in case, and
+                // to make TypeScript happy, fall back to -1 (which is a
+                // conventional exit status used for "exited with signal").
+                status: status ?? -1,
+              })
+            );
+          } else {
+            resolve();
+          }
+        });
+      });
+
+      try {
+        await completed;
+      } finally {
+        if (stdoutReplay !== undefined) {
+          await closeWriteStream(stdoutReplay);
+        }
+        if (stderrReplay !== undefined) {
+          await closeWriteStream(stderrReplay);
+        }
       }
-      if (stderrReplay !== undefined) {
-        await closeWriteStream(stderrReplay);
-      }
-    }
+    });
 
     this.#logger.log({
       script: this.#script,